Simplify support for multiple cluster types
* Renamed cloud_provider config to cluster_type
* cluster_type supports ec2 & existing clusters
* Ec2Cluster & ExistingCluster are now in seperates classes
* Ec2Cluster extends ExistingCluster and depends on boto
* ExistingCluster has now boto dependency
